fix(test) mock issue partially resolved, one false fail but to unit tests pass. I will revists
